move assets types into os
