+++ src/os/lib/markdown/markdownInput.tsx
@@ -1,4 +1,4 @@
-import { HasAssets } from "@/tools/essay/assets";
+import { HasAssets } from "@/os/assets";
 import { DocHandle } from "@automerge/automerge-repo";
 import { EditorView } from "@codemirror/view";
 import { useEffect, useState } from "react";
