+++ src/os/lib/markdown/useMarkdownPlugins.ts
@@ -16,7 +16,7 @@ import { searchKeymap } from "@codemirror/search";
 import { codeMonospacePlugin } from "./codemirrorPlugins/codeMonospace";
 import { lineWrappingPlugin } from "./codemirrorPlugins/lineWrapping";
 
-import { AssetsDoc, HasAssets } from "@/tools/essay/assets";
+import { AssetsDoc, HasAssets } from "@/os/assets";
 import { DocHandle, Repo } from "@automerge/automerge-repo";
 import { useRepo } from "@automerge/automerge-repo-react-hooks";
 import { useMemo } from "react";
